+++ src/interfaces/http.rs
@@ -219,7 +219,7 @@ impl Interface for HttpSocketInterface {
         use crate::interfaces::http::http_client::start_sending;
 
         let mut url = self.host.clone();
-        url.set_path(format!("{}{}{}", BASE_PATH, "/", NOTIFICATION_PATH).as_str());
+        url.set_path(format!("{}/{}", BASE_PATH, NOTIFICATION_PATH).as_str());
         debug!(target: LIB_LOG_TARGET, "Sending notification to: {}", url.as_str());
 
         tokio::spawn(async move { start_sending(interface_rx, shutdown, url.as_str()).await });