+++ src/endpoints/email.rs
@@ -18,7 +18,7 @@ use crate::endpoints::{Endpoint, EndpointConfig};
 use crate::notifications::{Key, ValidatedNotification};
 use crate::{Error, LIB_LOG_TARGET};
 use async_trait::async_trait;
-use log::{debug, error, info};
+use tracing::{debug, error, info};
 use mail_send::mail_builder::MessageBuilder;
 use mail_send::SmtpClientBuilder;
 use serde::Deserialize;