+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# v0.16.0
 ## Breaking Changes
 - remove vendored-tls feature